requirements/requirements-all.txt
black==23.11.0 ; python_version >= "3.10" and python_version < "3.12"
build==1.0.3 ; python_version >= "3.10" and python_version < "3.12"
cachecontrol[filecache]==0.13.1 ; python_version >= "3.10" and python_version < "3.12"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "3.12" and (sys_platform == "darwin" or sys_platform == "linux")
cfgv==3.4.0 ; python_version >= "3.10" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "3.12"
cleo==2.1.0 ; python_version >= "3.10" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.12"
contourpy==1.2.0 ; python_version >= "3.10" and python_version < "3.12"
crashtest==0.4.1 ; python_version >= "3.10" and python_version < "3.12"
cryptography==41.0.7 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "linux"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "3.12"
distlib==0.3.7 ; python_version >= "3.10" and python_version < "3.12"
dulwich==0.21.6 ; python_version >= "3.10" and python_version < "3.12"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
fastjsonschema==2.19.0 ; python_version >= "3.10" and python_version < "3.12"
filelock==3.13.1 ; python_version >= "3.10" and python_version < "3.12"
fonttools==4.45.1 ; python_version >= "3.10" and python_version < "3.12"
h5py==3.10.0 ; python_version >= "3.10" and python_version < "3.12"
identify==2.5.32 ; python_version >= "3.10" and python_version < "3.12"
idna==3.6 ; python_version >= "3.10" and python_version < "3.12"
importlib-metadata==7.0.0 ; python_version >= "3.10" and python_version < "3.12"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "3.12"
installer==0.7.0 ; python_version >= "3.10" and python_version < "3.12"
jaraco-classes==3.3.0 ; python_version >= "3.10" and python_version < "3.12"
jeepney==0.8.0 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "linux"
keyring==24.3.0 ; python_version >= "3.10" and python_version < "3.12"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "3.12"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.12"
matplotlib==3.8.2 ; python_version >= "3.10" and python_version < "3.12"
more-itertools==10.1.0 ; python_version >= "3.10" and python_version < "3.12"
mpi4py==3.1.5 ; python_version >= "3.10" and python_version < "3.12"
msgpack==1.0.7 ; python_version >= "3.10" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.12"
mypy==1.7.1 ; python_version >= "3.10" and python_version < "3.12"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "3.12"
numpy==1.25.2 ; python_version >= "3.10" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.12"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "3.12"
pexpect==4.9.0 ; python_version >= "3.10" and python_version < "3.12"
pillow==10.1.0 ; python_version >= "3.10" and python_version < "3.12"
pkginfo==1.9.6 ; python_version >= "3.10" and python_version < "3.12"
platformdirs==3.11.0 ; python_version >= "3.10" and python_version < "3.12"
pluggy==1.3.0 ; python_version >= "3.10" and python_version < "3.12"
pre-commit==3.5.0 ; python_version >= "3.10" and python_version < "3.12"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.10" and python_version < "3.12" and (sys_platform == "darwin" or sys_platform == "linux")
pyparsing==3.1.1 ; python_version >= "3.10" and python_version < "3.12"
pyproject-hooks==1.0.0 ; python_version >= "3.10" and python_version < "3.12"
pytest==7.4.3 ; python_version >= "3.10" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.12"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.12"
pywin32-ctypes==0.2.2 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "win32"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.12"
rapidfuzz==3.5.2 ; python_version >= "3.10" and python_version < "3.12"
rich==14.2.0 ; python_version >= "3.10" and python_version < "3.12"
ruff==0.0.261 ; python_version >= "3.10" and python_version < "3.12"
scipy==1.9.3 ; python_version >= "3.10" and python_version < "3.12"
secretstorage==3.3.3 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "linux"
setuptools==69.0.2 ; python_version >= "3.10" and python_version < "3.12"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.12"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.12.3 ; python_version >= "3.10" and python_version < "3.12"
trove-classifiers==2023.11.29 ; python_version >= "3.10" and python_version < "3.12"
types-requests==2.31.0.6 ; python_version >= "3.10" and python_version < "3.12"
types-urllib3==1.26.25.14 ; python_version >= "3.10" and python_version < "3.12"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "3.12"
urllib3==1.26.18 ; python_version >= "3.10" and python_version < "3.12"
virtualenv==20.24.7 ; python_version >= "3.10" and python_version < "3.12"
webdav4==0.10.0 ; python_version >= "3.10" and python_version < "3.12"
xattr==0.10.1 ; python_version >= "3.10" and python_version < "3.12" and sys_platform == "darwin"
zipp==3.17.0 ; python_version >= "3.10" and python_version < "3.12"
